==> data collection/00 - Initial data collection.ipynb <==
import pandas as pd
import copy
import numpy as np

<h3>Import file with RUNTS subscribers </h3>

file_path='Iscritti_20112024.xlsx'
df = pd.read_excel(file_path)

df

# Counts of the different provinces
conteggio_province = df['Provincia Sede legale\r\nProvinz Rechtssitz'].value_counts()
conteggio_province

# Dictionary that maps provinces to regions
province_to_regione = {
    'AG': 'Sicilia', 'AL': 'Piemonte', 'AN': 'Marche', 'AO': 'Valle d\'Aosta', 'AR': 'Toscana',
    'AP': 'Marche', 'AT': 'Piemonte', 'AV': 'Campania', 'BA': 'Puglia', 'BT': 'Puglia',
    'BL': 'Veneto', 'BN': 'Campania', 'BG': 'Lombardia', 'BI': 'Piemonte', 'BO': 'Emilia-Romagna',
    'BZ': 'Trentino-Alto Adige', 'BS': 'Lombardia', 'BR': 'Puglia', 'CA': 'Sardegna', 'CL': 'Sicilia',
    'CB': 'Molise', 'CI': 'Sardegna', 'CE': 'Campania', 'CH': 'Abruzzo', 'CO': 'Lombardia',
    'CS': 'Calabria', 'CR': 'Lombardia', 'KR': 'Calabria', 'CN': 'Piemonte', 'EN': 'Sicilia',
    'FM': 'Marche', 'FE': 'Emilia-Romagna', 'FI': 'Toscana', 'FG': 'Puglia', 'FC': 'Emilia-Romagna',
    'FR': 'Lazio', 'GE': 'Liguria', 'GO': 'Friuli-Venezia Giulia', 'GR': 'Toscana', 'IM': 'Liguria',
    'IS': 'Molise', 'SP': 'Liguria', 'LT': 'Lazio', 'LE': 'Puglia', 'LC': 'Lombardia',
    'LI': 'Toscana', 'LO': 'Lombardia', 'LU': 'Toscana', 'MC': 'Marche', 'MN': 'Lombardia',
    'MS': 'Toscana', 'MT': 'Basilicata', 'ME': 'Sicilia', 'MI': 'Lombardia', 'MO': 'Emilia-Romagna',
    'MB': 'Lombardia', 'NA': 'Campania', 'NO': 'Piemonte', 'NU': 'Sardegna', 'OR': 'Sardegna',
    'PD': 'Veneto', 'PA': 'Sicilia', 'PR': 'Emilia-Romagna', 'PV': 'Lombardia', 'PG': 'Umbria',
    'PS': 'Marche', 'PE': 'Abruzzo', 'PC': 'Emilia-Romagna', 'PI': 'Toscana', 'PT': 'Toscana',
    'PN': 'Friuli-Venezia Giulia', 'PZ': 'Basilicata', 'PO': 'Toscana', 'RG': 'Sicilia', 'RA': 'Emilia-Romagna',
    'RC': 'Calabria', 'RE': 'Emilia-Romagna', 'RI': 'Lazio', 'RN': 'Emilia-Romagna', 'RM': 'Lazio',
    'RO': 'Veneto', 'SA': 'Campania', 'SS': 'Sardegna', 'SV': 'Liguria', 'SI': 'Toscana',
    'SR': 'Sicilia', 'SO': 'Lombardia', 'TA': 'Puglia', 'TE': 'Abruzzo', 'TR': 'Umbria',
    'TO': 'Piemonte', 'TP': 'Sicilia', 'TN': 'Trentino-Alto Adige', 'TV': 'Veneto', 'TS': 'Friuli-Venezia Giulia',
    'UD': 'Friuli-Venezia Giulia', 'VA': 'Lombardia', 'VE': 'Veneto', 'VB': 'Piemonte', 'VC': 'Piemonte',
    'VR': 'Veneto', 'VV': 'Calabria', 'VI': 'Veneto', 'VT': 'Lazio'
}

# Create a new column with the Region mapping
df['Regione'] = df['Provincia Sede legale\r\nProvinz Rechtssitz'].map(province_to_regione)


df

# Counts of the different regions
conteggio_regione = df['Regione'].value_counts()
conteggio_regione

# Counts the different types of entities - These will be the labels of the graph's nodes
df['Sezione (**)\r\nSektion'].value_counts()

<h3>Build the Lombardia dataframe</h3>

df_lombardia = df[df['Regione'] == 'Lombardia'].reset_index(drop=True)
df_lombardia

df_lombardia.columns

# Change the names (in order to drop the translation)
df_lombardia.columns = ['Codice fiscale', 'Repertorio', 'Denominazione', 'Sezione', 'Cognome rapp legale', 'Nome rapp legale', 'Rete', 'Comune', 'Provincia', '5x1000', 'Data iscrizione', 'Regione'] 


df_lombardia

# Save the df in a csv
df_lombardia.to_csv('df_lombardia.csv', index=False) 


lista_denominazioni = df_lombardia['Denominazione'].tolist()
lista_denominazioni
